fixes according to comments
