+++ 2-patterns_integrations/src/main/java/com/epam/dhontar/aqamp/utils/enums/TestRailCredentials.java
@@ -1,13 +1,13 @@
 package com.epam.dhontar.aqamp.utils.enums;
 
-public enum Constants {
+public enum TestRailCredentials {
     TESTRAIL_USER("dimahtest@gmail.com"),
     TESTRAIL_PWD("?pKSTr8ApHo8MJGk"),
     TESTRAIL_URL("https://aqadmitry.testrail.io/");
 
     private final String constantValue;
 
-    Constants(String constantValue) {
+    TestRailCredentials(String constantValue) {
         this.constantValue = constantValue;
     }
