+++ 2-patterns_integrations/src/main/java/com/epam/dhontar/aqamp/utils/integrations/slack/entity/SlackMessage.java
@@ -10,10 +10,6 @@ public class SlackMessage {
     public SlackMessage() {
     }
 
-    public void setText(String text) {
-        this.text = text;
-    }
-
     public String getText() {
         return text;
     }
