+++ 2-patterns_integrations/testng.xml
@@ -3,7 +3,8 @@
 <suite name="All tests">
     <test name="Test run over users and authors endpoints">
          <classes>
-             <class name="com.epam.dhontar.aqamp.RestTests"/>
+             <class name="com.epam.dhontar.aqamp.UsersTests"/>
+             <class name="com.epam.dhontar.aqamp.AuthorsTests"/>
           </classes>
      </test>
 </suite>
